Moved routines to add and remove admins from the command to Permissions
